GCE/PubSub/requirements.txt
google-auth==2.11.1
google-cloud-pubsub==2.13.10
tweepy==4.10.1